refactor: restructured tests
