+++ src/lib/client.ts
@@ -98,14 +98,14 @@ export interface EncryptDataPayload {
   keySource: KeySource;
   plaintext: Uint8Array;
   encryptMetadata?: EncryptMetadata;
-  metadata: Metadata;
+  metadata?: Metadata;
 }
 export interface DecryptDataPayload {
   profile: string;
   keySource: KeySource;
   ciphertext: Uint8Array;
   decryptMetadata?: DecryptMetadata;
-  metadata: Metadata;
+  metadata?: Metadata;
 }
 
 const breakers = new WeakMap<object, Map<string, CircuitBreaker>>();
